[REFACTOR] Split severity api test
